refactor: rename file to view diff
