+++ packages/gatsby-theme-docs/src/components/image.js
